last minute changes denk ik

+++ teams/views.py
@@ -1,16 +1,20 @@
 from django.shortcuts import render, get_object_or_404
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth import login, authenticate
+from django.contrib.auth.decorators import user_passes_test
+from django.core.mail import send_mail
 from django.core.urlresolvers import reverse, reverse_lazy
 from django.http import HttpResponseRedirect
-from teams.models import Team, TemporaryUser, RiallyUser
-from teams.forms import TemporaryUserForm, TemporaryUserConfirmForm
+from django.template.loader import render_to_string
+from django.utils.decorators import method_decorator
 from django.views.generic.edit import CreateView
 from django.views.generic.list import ListView
 from django.views.generic.detail import DetailView
-from django.core.mail import send_mail
-from django.template.loader import render_to_string
-from django.contrib.auth.forms import UserCreationForm
 
-from rially.decorators import require_team_captain
+from teams.models import Team, TemporaryUser, RiallyUser
+from teams.forms import TemporaryUserForm, TemporaryUserConfirmForm, TeamCreateForm
+from rially.decorators import require_team_captain, require_riallyuser
+
 import random
 import string
 
@@ -19,13 +23,17 @@ def my_team(request):
 
 @require_team_captain(login_url="/login")
 def delete(request, teamId, userId):
-    team = get_object_or_404(Team, pk=teamId)
 
-    riallyuser = get_object_or_404(team.riallyuser_set, pk=userId)
-    riallyuser.user.delete()
-    riallyuser.delete()
+    if (request.method == 'POST'):
+        team = get_object_or_404(Team, pk=teamId)
 
-    return HttpResponseRedirect(reverse('teams.views.my_team'))
+        riallyuser = get_object_or_404(team.riallyuser_set, pk=userId)
+        riallyuser.user.delete()
+        riallyuser.delete()
+
+        return HttpResponseRedirect(reverse('teams.views.my_team'))
+    else:
+        return render(request, 'team_user_delete.html')
 
 def team_confirm(request, slug):
     temp_user = get_object_or_404(TemporaryUser, email_link=slug)
@@ -42,6 +50,10 @@ def team_confirm(request, slug):
             # maak dan een nieuwe rially_user voor deze user
             rially_user = RiallyUser(user=instance, team=temp_user.team, is_team_captain=temp_user.becomes_team_captain)
             rially_user.save()
+
+            # voor de handig, log deze persoon ook in
+            instance.backend = 'django.contrib.auth.backends.ModelBackend'
+            login(request, instance)
 
             # verwijder dan de temp_user zodat men zich niet twee keer aanmeld
             temp_user.delete()
@@ -62,6 +74,8 @@ class TemporaryUserCreateView(CreateView):
     fields = ['email']
 
     def form_valid(self, form):
+        if not self.request.user.riallyuser.is_team_captain:
+            raise ValidationError("Je kan geen users toevoegen als je geen admin bent")
         form.instance.team = self.request.user.riallyuser.team
         form.instance.email_link = ''.join(random.choice(string.ascii_lowercase + string.ascii_uppercase + string.digits) for x in range(64))
         form.save()
@@ -73,9 +87,11 @@ class TemporaryUserCreateView(CreateView):
         return super(TemporaryUserCreateView, self).form_valid(form)
 
     def get_success_url(self):
-        team = self.request.user.riallyuser.team
-
         return reverse('teams.views.my_team')
+
+    @method_decorator(require_riallyuser)
+    def dispatch(self, *args, **kwargs):
+        return super(TemporaryUserCreateView, self).dispatch(*args, **kwargs)
 
 class TeamView(ListView):
     template_name = 'team_list.html'
@@ -86,3 +102,30 @@ class TeamDetail(DetailView):
     model = Team
     template_name = 'team_detail.html'
     context_object_name = 'team'
+
+class TeamCreateView(CreateView):
+    model = Team
+    template_name = 'team_create.html'
+    form_class = TeamCreateForm
+
+    def form_valid(self, form):
+        instance = form.save()
+
+        email_link = ''.join(random.choice(string.ascii_lowercase + string.ascii_uppercase + string.digits) for x in range(64))
+
+        temp_user = TemporaryUser(team=instance, email=form.cleaned_data["email"], email_link=email_link, becomes_team_captain=True)
+        temp_user.save()
+
+        email = render_to_string('new_team_email.txt', {'team': form.instance, 'email_link': temp_user.get_absolute_url()})
+
+        send_mail('Toegevoegd aan een Rially team!', email, 'rially@inter-actief.net',
+    [form.cleaned_data["email"]], fail_silently=False)
+
+        return super(TeamCreateView, self).form_valid(form)
+
+    def get_success_url(self):
+        return reverse('all_teams_view')
+
+    @method_decorator(user_passes_test(lambda u: u.is_superuser, login_url="/login/"))
+    def dispatch(self, *args, **kwargs):
+        return super(TeamCreateView, self).dispatch(*args, **kwargs)
